refactor: de-part victory overlay and extract panel bus hosts (#4117)
Convert victory_overlay panel/actions/corners from part libraries to
explicit-import modules (shrink repo.app_no_part_directives grandfather by
four entries). Move PanelBusDialogHost scaffolding to app/test/
panel_bus_dialog_host.dart outside the support LOC tree and ratchet
appTestSupportLocCeiling 8302 → 8142.

Refs #4117
