+++ app/lib/features/game/flame/overlays/victory_overlay_panel.dart
@@ -1,4 +1,17 @@
-part of 'victory_overlay.dart';
+import 'package:flutter/material.dart';
+import 'package:colonizethis_app_l10n/l10n/l10n.dart';
+
+import 'package:colonizethis_logic/colonizethis_logic.dart';
+import 'package:colonizethis_models/colonizethis_models.dart' as ct_models;
+
+import '../../../../config/constants.dart';
+import 'package:colonizethis_app_ui_chrome/config/editorial_monocle_palette.dart';
+import 'package:colonizethis_app_ui_chrome/widgets/ct_brass_divider.dart';
+import '../../../../widgets/ct_gradients.dart';
+import '../../../../widgets/ct_spacing.dart';
+
+import 'victory_overlay_panel_actions.dart';
+import 'victory_overlay_panel_corners.dart';
 
 /// Brass-bordered ceremonial panel for the military victory overlay.
 ///
@@ -58,12 +71,12 @@ class VictoryPanel extends StatelessWidget {
             const Positioned(
               top: cornerBracketInset,
               left: cornerBracketInset,
-              child: _VictoryCornerBracket(corner: _CornerSide.topLeft),
+              child: VictoryCornerBracket.topLeft(),
             ),
             const Positioned(
               bottom: cornerBracketInset,
               right: cornerBracketInset,
-              child: _VictoryCornerBracket(corner: _CornerSide.bottomRight),
+              child: VictoryCornerBracket.bottomRight(),
             ),
           ],
         ),
@@ -110,7 +123,7 @@ class VictoryPanel extends StatelessWidget {
     return Column(
       mainAxisSize: MainAxisSize.min,
       children: <Widget>[
-        _VictoryLaurelRow(narrow: narrow),
+        VictoryLaurelRow(narrow: narrow),
         const SizedBox(height: 10),
         Text(
           victoryLabel.toUpperCase(),
@@ -126,7 +139,7 @@ class VictoryPanel extends StatelessWidget {
           textAlign: TextAlign.center,
         ),
         const SizedBox(height: 24),
-        buildActionRow(context, l10n, narrow: narrow),
+        buildVictoryPanelActionRow(this, context, l10n, narrow: narrow),
       ],
     );
   }
