+++ app/lib/features/game/flame/overlays/victory_overlay_panel_corners.dart
@@ -1,14 +1,24 @@
-part of 'victory_overlay.dart';
+import 'package:flutter/material.dart';
 
-enum _CornerSide { topLeft, bottomRight }
+import 'package:colonizethis_app_ui_chrome/config/editorial_monocle_palette.dart';
+
+import 'victory_overlay_panel.dart';
+
+enum VictoryCornerSide { topLeft, bottomRight }
 
 /// Asymmetric corner-bracket ornament drawn for the top-left and
 /// bottom-right corners of the victory panel surface. Renders a 1.5px brass
 /// L-shape at `cornerBracketAlpha` opacity.
-class _VictoryCornerBracket extends StatelessWidget {
-  const _VictoryCornerBracket({required this.corner});
+class VictoryCornerBracket extends StatelessWidget {
+  const VictoryCornerBracket({required this.corner, super.key});
 
-  final _CornerSide corner;
+  final VictoryCornerSide corner;
+
+  const VictoryCornerBracket.topLeft({super.key})
+      : corner = VictoryCornerSide.topLeft;
+
+  const VictoryCornerBracket.bottomRight({super.key})
+      : corner = VictoryCornerSide.bottomRight;
 
   @override
   Widget build(BuildContext context) {
@@ -37,7 +47,7 @@ class _VictoryCornerBracketPainter extends CustomPainter {
 
   final Color color;
   final double stroke;
-  final _CornerSide corner;
+  final VictoryCornerSide corner;
 
   @override
   void paint(Canvas canvas, Size size) {
@@ -48,7 +58,7 @@ class _VictoryCornerBracketPainter extends CustomPainter {
 
     final double half = stroke / 2;
     switch (corner) {
-      case _CornerSide.topLeft:
+      case VictoryCornerSide.topLeft:
         canvas.drawLine(
           Offset(0, half),
           Offset(size.width, half),
@@ -60,7 +70,7 @@ class _VictoryCornerBracketPainter extends CustomPainter {
           paint,
         );
         break;
-      case _CornerSide.bottomRight:
+      case VictoryCornerSide.bottomRight:
         canvas.drawLine(
           Offset(0, size.height - half),
           Offset(size.width, size.height - half),
