+++ tool/app_no_part_directives_grandfather.dart
@@ -44,10 +44,6 @@ const List<String> appNoPartDirectivesGrandfatheredForTests = <String>[
   'app/lib/features/game/flame/map_state/game_map_area_view.dart',
   'app/lib/features/game/flame/overlays/debug_console_overlay_panel.dart',
   'app/lib/features/game/flame/overlays/debug_console_overlay_panel_state.dart',
-  'app/lib/features/game/flame/overlays/victory_overlay.dart',
-  'app/lib/features/game/flame/overlays/victory_overlay_panel.dart',
-  'app/lib/features/game/flame/overlays/victory_overlay_panel_actions.dart',
-  'app/lib/features/game/flame/overlays/victory_overlay_panel_corners.dart',
   'app/lib/features/game/flame/tilesets/terrain_tileset.dart',
   'app/lib/features/game/flame/tilesets/terrain_tileset_loaders.dart',
   'app/lib/features/game/flame/tilesets/terrain_tileset_models.dart',
